Space rightside children, first draft
The children on the rightside are spaced somewhat well, but I expect that
it isn't perfefct and I will have to make adjustments later.

Everything having to do with the form is not correct.
That will be the next thing to tackle.

Log in link will be created eventually, last priority along with font
implementation.